requirements.txt
psycopg[binary]
torch
python-dotenv
pgvector
Pillow
tqdm
transformers
